+++ sdk/containerinstance/azure-mgmt-containerinstance/generated_samples/subnet_service_association_link_delete.py
@@ -7,6 +7,7 @@
 # --------------------------------------------------------------------------
 
 from azure.identity import DefaultAzureCredential
+
 from azure.mgmt.containerinstance import ContainerInstanceManagementClient
 
 """
@@ -29,14 +30,13 @@ def main():
         subscription_id="subid",
     )
 
-    response = client.subnet_service_association_link.begin_delete(
+    client.subnet_service_association_link.begin_delete(
         resource_group_name="demo",
         virtual_network_name="demo2",
         subnet_name="demo3",
     ).result()
-    print(response)
 
 
-# x-ms-original-file: specification/containerinstance/resource-manager/Microsoft.ContainerInstance/stable/2023-05-01/examples/SubnetServiceAssociationLinkDelete.json
+# x-ms-original-file: specification/containerinstance/resource-manager/Microsoft.ContainerInstance/preview/2024-10-01-preview/examples/SubnetServiceAssociationLinkDelete.json
 if __name__ == "__main__":
     main()